test: verify listOf & setOf have same modifiability
Add tests that verify that the default listOf() and setOf()
produce containers with the same modifiability. That is, whether
listOf() produces a modifiable list or an unmodifiable list, setOf()
should produce a set with the same (un)modifiability.

There are already tests that show that emptyList() and emptySet()
perform the same as listOf() and setOf().