fix(cal): handle when Actual is not a Calendar
After correctly detecting that the type of the actual item is not a
Calendar, also produce a sensible and useful diagnostic description,
rather than crashing with a class cast exception.
This error was rare because we usually rely on the matcher's client to
ensure the class of the actual, for example Hamcrest's assetThat()
will not compile if the actual is not a Calendar, to match the Calendar
specified by our Matcher's signature.