build: fix Spotless: 2 spaces for Markdown/Text
Although correctly specifying spaces rather than tabs for Java,
Markdown, and for Text, it was also specifying 4 spacesPerTab.
This latter is not used to _check_ the indentation (so far as I can
see), but it _is_ used to _correct_ any tabs found, when using
`spotless:apply`. But our standard for Markdown is _2_ spaces, not 4,
and we follow that with Text files too (partially because we sometimes
histoircally used `.md.txt` as a descriptive extension, for people who
don't understand Markdown .md extension.
So create a separate format entry for Markdown and Text, specifying 2
spaces.

Matches the ayeseeem.org style guide at
https://github.com/ayeseeem/icm-java-style